# Conflicts:
#	gangwonActivity/src/main/java/com/multicampus/gangwonActivity/common/ResponseCode.java
#	gangwonActivity/src/main/java/com/multicampus/gangwonActivity/common/ResponseMessage.java
#	gangwonActivity/src/main/java/com/multicampus/gangwonActivity/controller/MyPageController.java
#	gangwonActivity/src/main/java/com/multicampus/gangwonActivity/service/MyPageService.java
#	gangwonActivity/src/main/java/com/multicampus/gangwonActivity/service/implement/MyPageServiceImpl.java

+++ gangwonActivity/src/main/java/com/multicampus/gangwonActivity/service/implement/MyPageServiceImpl.java
@@ -1,5 +1,6 @@
 package com.multicampus.gangwonActivity.service.implement;
 
+import com.multicampus.gangwonActivity.dto.request.mypage.CheckPasswordRequestDto;
 import com.multicampus.gangwonActivity.dto.request.mypage.ModifyMyInfoRequestDto;
 import com.multicampus.gangwonActivity.dto.response.ResponseDto;
 import com.multicampus.gangwonActivity.dto.response.board.GetBoardListResponseDto;
@@ -86,13 +87,13 @@ public class MyPageServiceImpl implements MyPageService {
                 return ModifyMyInfoResponseDto.notExistUser();
             }
 
-            if(modifyPassword != null) {
+            if(modifyPassword != null && modifyPassword !="") {
                 String encodedChangePassword = passwordEncoder.encode(modifyPassword);
                 user.ModifyPassword(encodedChangePassword);
                 userRepository.save(user);
             }
 
-            if(modifyUserNick != null){
+            if(modifyUserNick != null && modifyUserNick !=""){
                 user.ModifyUserNick(modifyUserNick);
                 userRepository.save(user);
             }
@@ -106,13 +107,16 @@ public class MyPageServiceImpl implements MyPageService {
 
 
     @Override
-    public ResponseEntity<? super ModifyMyInfoResponseDto> deleteMyInfo(String id) {
+    public ResponseEntity<? super ModifyMyInfoResponseDto> deleteMyInfo(String id, ModifyMyInfoRequestDto dto) {
         try {
             User user = userRepository.findByUserId(id);
             if(user == null) {
                 return ModifyMyInfoResponseDto.notExistUser();
             }
+            String userId = dto.getUserId();
+            String UserId = user.getUserId();
 
+            if(!UserId.equals(userId)) return ModifyMyInfoResponseDto.idError();
             userRepository.delete(user);
 
         } catch (Exception e) {
@@ -130,6 +134,23 @@ public class MyPageServiceImpl implements MyPageService {
 
         return adminMapper.userInfo(userNo);
     }
+
+    @Override
+    public ResponseEntity<? super ModifyMyInfoResponseDto> checkPassword(String id, CheckPasswordRequestDto dto) {
+        try {
+        User user =userRepository.findByUserId(id);
+        if (user==null) return ModifyMyInfoResponseDto.notExistUser();
+
+        boolean isMatch = passwordEncoder.matches(dto.getUserPassword(), user.getUserPassword());
+        if(!isMatch) return ModifyMyInfoResponseDto.passwordInFailed();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseDto.databaseError();
+        }
+
+        return ModifyMyInfoResponseDto.success();
+    }
+
 
     // 내 추천 코스 ( 민호형 & 수지 )
     @Override
